backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlite-utils==3.30
pydantic==1.10.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.28.0
torch==2.0.0
ctransformers==0.2.8
spacy==3.5.0
